# docs/sms-api-doc.yaml
openapi: 3.0.1
info:
  title: School Management System API
  version: "1.3.3"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /config/{key}:
    get:
      tags:
      - App Config
      summary: Returns configuration by key
      description: "Retrieves single Config data based on Config key. Required authority\
        \ level: Administrator"
      operationId: getConfigByKey
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Config with given key not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AppConfigDetailedResponse'
        "200":
          description: Config value successfully fetched
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AppConfigDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
    patch:
      tags:
      - App Config
      summary: Update configuration value
      description: "Modifies one of global application parameters and settings. Required\
        \ authority level: Administrator"
      operationId: updateConfig
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppConfigUpdateRequest'
        required: true
      responses:
        "200":
          description: Config value successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AppConfigDetailedResponse'
        "404":
          description: Config with given key not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AppConfigDetailedResponse'
        "406":
          description: Config is uneditable
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AppConfigDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
...
  /config:
    get:
      tags:
      - App Config
      summary: List editable system configs
      description: "Retrieves all global configuration data that is editable for the\
        \ User. Required authority level: Administrator"
      operationId: getAllEditableConfigs
      responses:
        "401":
          description: Full authentication is required to access this resource
        "200":
          description: Returns config values (empty result as well)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppConfigSummaryResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    AuthorityEntity:
      type: object
      properties:
        id:
          type: integer
          format: int32
        authority:
          type: string
...
    AppConfigUpdateRequest:
      type: object
      properties:
        value:
          maxLength: 150
          minLength: 0
          type: string
        description:
          maxLength: 200
          minLength: 0
          type: string
        authoritiesToAddAsEditableBy:
          type: array
          items:
            type: integer
            format: int32
        authoritiesToRemoveAsEditableBy:
          type: array
          items:
            type: integer
            format: int32
        minValue:
          maxLength: 30
          minLength: 0
          type: string
        maxValue:
          maxLength: 30
          minLength: 0
          type: string
    AppConfigDetailedResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
        value:
          type: string
        valueType:
          type: string
          enum:
          - INTEGER
          - TEXT
          - BOOLEAN
        description:
          type: string
        editableBy:
          type: array
          items:
            $ref: '#/components/schemas/AuthorityEntity'
        minValue:
          type: string
        maxValue:
          type: string
        createdAt:
          type: string
          format: date-time
        modifiedByUserId:
          type: integer
          format: int32
        modifiedAt:
          type: string
          format: date-time
        editable:
          type: boolean
...
    AppConfigSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
        value:
          type: string
        editableBy:
          type: array
          items:
            $ref: '#/components/schemas/AuthorityEntity'
        description:
          type: string
        editable:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT